refactor(config): refactor unet/ to use pydantic config
