+++ src/autovisionai/models/unet/unet_trainer.py
@@ -6,7 +6,7 @@ import torch.nn as nn
 from torch.optim import Adam, Optimizer
 from torch.optim.lr_scheduler import StepLR
 
-from autovisionai.configs.config import CONFIG
+from autovisionai.configs import CONFIG, LRSchedulerConfig, OptimizerConfig
 from autovisionai.loggers.ml_logging import log_image_to_all_loggers
 from autovisionai.models.unet.unet_model import Unet
 from autovisionai.utils.utils import get_batch_images_and_pred_masks_in_a_grid, masks_iou
@@ -89,16 +89,19 @@ class UnetTrainer(pl.LightningModule):
         self.val_outputs.clear()
 
     def configure_optimizers(self) -> Dict[str, Union[Optimizer, object]]:
+        optim_cfg: OptimizerConfig = CONFIG.models.unet.optimizer
+        lr_scheduler_cfg: LRSchedulerConfig = CONFIG.models.unet.lr_scheduler
+
         optimizer = Adam(
             self.model.parameters(),
-            lr=CONFIG["unet"]["optimizer"]["initial_lr"].get(),
-            weight_decay=CONFIG["unet"]["optimizer"]["weight_decay"].get(),
+            lr=optim_cfg.initial_lr,
+            weight_decay=optim_cfg.weight_decay,
         )
 
         lr_scheduler = StepLR(
             optimizer,
-            step_size=CONFIG["unet"]["lr_scheduler"]["step_size"].get(),
-            gamma=CONFIG["unet"]["lr_scheduler"]["gamma"].get(),
+            step_size=lr_scheduler_cfg.step_size,
+            gamma=lr_scheduler_cfg.gamma,
         )
 
         return {"optimizer": optimizer, "lr_scheduler": lr_scheduler}
